+++ lib/src/generator/properties_generator_mixin.dart
@@ -3,9 +3,11 @@
 // ignore_for_file: use_raw_strings
 // ignore_for_file: unnecessary_raw_strings
 
+import 'package:meta/meta.dart';
 import 'package:pubspec_generator/src/generator/pubspec_generator.dart';
 
-/// {@nodoc}
+@internal
+@immutable
 mixin PropertiesGeneratorMixin on PubspecGenerator {
   @override
   Iterable<String> generate(Map<String, Object> pubspec) sync* {