+++ lib/src/generator/source_generator_mixin.dart
@@ -1,18 +1,20 @@
 // ignore_for_file: avoid_escaping_inner_quotes, avoid_annotating_with_dynamic
 
+import 'package:meta/meta.dart';
 import 'package:pubspec_generator/src/generator/pubspec_generator.dart';
 import 'package:pubspec_generator/src/generator/representation.dart';
 
-/// {@nodoc}
+@internal
+@immutable
 mixin SourceGeneratorMixin on PubspecGenerator {
   @override
-  Iterable<String> generate(Map<String, Object> data) sync* {
+  Iterable<String> generate(Map<String, Object> pubspec) sync* {
     final buffer = StringBuffer()
       ..writeln('/// Source data from pubspec.yaml')
       ..writeln('static const Map<String, Object> source = <String, Object>{');
 
     void refToValue(String key, String value) {
-      data.remove(key);
+      pubspec.remove(key);
       buffer
         ..write('  \'')
         ..write(key)
@@ -38,8 +40,8 @@ mixin SourceGeneratorMixin on PubspecGenerator {
     refToValue('dependencies', 'dependencies');
     refToValue('dev_dependencies', 'devDependencies');
     refToValue('dependency_overrides', 'dependencyOverrides');
-    if (data.isNotEmpty) {
-      for (final entry in data.entries) {
+    if (pubspec.isNotEmpty) {
+      for (final entry in pubspec.entries) {
         buffer
           ..write('  \'')
           ..write(entry.key)
@@ -54,6 +56,6 @@ mixin SourceGeneratorMixin on PubspecGenerator {
     }
     buffer.writeln('};');
     yield buffer.toString();
-    yield* super.generate(data);
+    yield* super.generate(pubspec);
   }
 }