+++ lib/src/generator/topics_generator_mixin.dart
@@ -1,9 +1,11 @@
 // ignore_for_file: avoid_escaping_inner_quotes, avoid_annotating_with_dynamic
 
+import 'package:meta/meta.dart';
 import 'package:pubspec_generator/src/generator/pubspec_generator.dart';
 import 'package:pubspec_generator/src/generator/representation.dart';
 
-/// {@nodoc}
+@internal
+@immutable
 mixin TopicsGeneratorMixin on PubspecGenerator {
   @override
   Iterable<String> generate(Map<String, Object> pubspec) sync* {
@@ -27,7 +29,9 @@ const String _$topicsDescription = '''
 ///
 /// Current app [topics]
 ///
-/// Package authors can use the topics field to categorize their package. Topics can be used to assist discoverability during search with filters on pub.dev. Pub.dev displays the topics on the package page as well as in the search results.
+/// Package authors can use the topics field to categorize their package.
+/// Topics can be used to assist discoverability during search with filters on pub.dev.
+/// Pub.dev displays the topics on the package page as well as in the search results.
 ///
 /// The field consists of a list of names. For example:
 ///