+++ lib/src/generator/version_generator_mixin.dart
@@ -1,10 +1,12 @@
 // ignore_for_file: avoid_escaping_inner_quotes, avoid_annotating_with_dynamic
 // ignore_for_file: unnecessary_raw_strings
 
+import 'package:meta/meta.dart';
 import 'package:pub_semver/pub_semver.dart' as ver;
 import 'package:pubspec_generator/src/generator/pubspec_generator.dart';
 
-/// {@nodoc}
+@internal
+@immutable
 mixin VersionGeneratorMixin on PubspecGenerator {
   @override
   Iterable<String> generate(Map<String, Object> pubspec) sync* {