+++ lib/src/pubspec_builder.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: avoid_escaping_inner_quotes, avoid_annotating_with_dynamic
 
 import 'package:build/build.dart';
-
+import 'package:meta/meta.dart';
 import 'package:pubspec_generator/src/generator/pubspec_generator.dart';
 import 'package:pubspec_generator/src/generator/pubspec_generator_impl.dart';
 import 'package:pubspec_generator/src/parser/pubspec_parser.dart';
@@ -14,6 +14,8 @@ Builder pubspecBuilder(BuilderOptions options) =>
     PubspecBuilderImpl(PubspecBuilderConfig.fromBuilderOptions(options));
 
 /// Builder
+@internal
+@immutable
 abstract class PubspecBuilder implements Builder {
   /// Pubspec parser
   final PubspecParser pubspecParser;
@@ -30,7 +32,7 @@ abstract class PubspecBuilder implements Builder {
   /// PubspecBuilder constructor with BuilderOptions
   PubspecBuilder(this.config)
       : pubspecParser = const PubspecParserImpl(),
-        pubspecGenerator = const PubspecGeneratorImpl(),
+        pubspecGenerator = PubspecGeneratorImpl(config: config),
         buildExtensions = <String, List<String>>{
           'pubspec.yaml': <String>[config.output],
         };